feat: bar chart
